+++ wl-extract/rsakey.h
@@ -25,7 +25,7 @@ typedef unsigned long mp_digit;
 
 struct mp_int
 {
-    int used, alloc, sign;
+    size_t used, alloc, sign;
     mp_digit* dp;
 };
 
@@ -59,7 +59,7 @@ public:
     ~rsa_public_key() noexcept { clear(); }
 
     bool set_from_mpints(const mp_int* mod, const mp_int* exp);
-    bool set_from_bytes(const uint8_t* mod_bytes, int mod_used, const uint8_t* exp_bytes, int exp_used);
+    bool set_from_bytes(const uint8_t* mod_bytes, size_t mod_used, const uint8_t* exp_bytes, size_t exp_used);
 
     void clear() noexcept;
